backend: working image upload

+++ backend/src/main/java/application/service/AnimalAdService.java
@@ -6,14 +6,17 @@ import application.dto.response.AnimalAdDto;
 import application.mapper.AnimalAdMapper;
 import application.model.AnimalAd;
 import application.model.Category;
+import application.model.Image;
 import application.model.User;
 import application.repository.AnimalAdRepository;
 import application.repository.CategoryRepository;
+import application.repository.ImageRepository;
 import application.repository.UserRepository;
 import application.security.JwtService;
 import lombok.Builder;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -27,6 +30,8 @@ public class AnimalAdService {
     private final UserRepository userRepository;
     private final AnimalAdMapper mapper;
     private final JwtService jwtService;
+    private final ImageService imageService;
+    private final ImageRepository imageRepository;
 
 
 
@@ -39,7 +44,7 @@ public class AnimalAdService {
         return mapper.toDto(ad);
     }
 
-    public AnimalAdDto createAd(CreateAdRequest request, String authHeader) {
+    public AnimalAdDto createAd(CreateAdRequest request, MultipartFile[] images, String authHeader) {
 
         String token = authHeader.substring(7);
 
@@ -62,16 +67,26 @@ public class AnimalAdService {
 
         AnimalAd saved=adRepository.save(ad);
 
+        imageService.saveImages(saved, images);
 
 
         return mapper.toDto(saved);
     }
 
 
-    public AnimalAdDto updateAd(Long id, UpdateAdRequest request) {
+    public AnimalAdDto updateAd(Long id, UpdateAdRequest request, MultipartFile[] newImages, List<Long> deleteImageIds, String authHeader) {
+
+        String token = authHeader.substring(7);
+
+        Long userId = jwtService.extractUserId(token);
 
         AnimalAd ad = adRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Ad not found"));
+
+        if (!ad.getUser().getId().equals(userId)) {
+            throw new RuntimeException("Not your ad");
+        }
+
 
         ad.setTitle(request.getTitle());
         ad.setDescription(request.getDescription());
@@ -85,11 +100,42 @@ public class AnimalAdService {
             ad.setCategory(category);
         }
 
+        if(deleteImageIds != null){
+
+            for (Long imageId : deleteImageIds) {
+
+                Image image = imageRepository.findById(imageId).orElseThrow(() -> new RuntimeException("image not found"));
+
+                imageService.deleteImage(image);
+
+            }
+
+        }
+
+        imageService.saveImages(ad, newImages);
+
         return mapper.toDto(adRepository.save(ad));
     }
 
 
-    public void deleteAd(Long id) {
+    public void deleteAd(Long id, String authHeader) {
+
+        String token = authHeader.substring(7);
+
+        Long userId = jwtService.extractUserId(token);
+
+        AnimalAd ad = adRepository.findById(id).orElseThrow(() -> new RuntimeException("ad not found"));
+
+        if (!ad.getUser().getId().equals(userId)) {
+            throw new RuntimeException("Not your ad");
+        }
+
+
+        for(Image image : ad.getImages()){
+            imageService.deleteImage(image);
+        }
+
+
         adRepository.deleteById(id);
     }
